hotfix : modify annotaion & indicator id type

+++ src/main/java/shinhan/intern/hotsignal/indicator/repository/IndicatorRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import shinhan.intern.hotsignal.indicator.entity.Indicator;
 
 @Repository
-public interface IndicatorRepository extends JpaRepository<Indicator, Integer> {
+public interface IndicatorRepository extends JpaRepository<Indicator, Long> {
     List<Indicator> findByCode(String code);
 }
